refactor: imports only from the index of actionCreators
