+++ src/core_modules/capture-core/components/DataEntries/Enrollment/epics/enrollment.epics.js
@@ -5,9 +5,8 @@ import { batchActionTypes, runRulesOnUpdateFieldBatch } from '../actions/enrollm
 import { actionTypes } from '../actions/enrollment.actions';
 import { getProgramFromProgramIdThrowIfNotFound, TrackerProgram } from '../../../../metaData';
 import { getCurrentClientValues, getCurrentClientMainData } from '../../../../rules/actionsCreator';
+import type { FieldData } from '../../../../rules/actionsCreator';
 import getDataEntryKey from '../../../DataEntry/common/getDataEntryKey';
-import type
-{ FieldData } from '../../../../rules/actionsCreator/inputHelpers';
 
 const errorMessages = {
     PROGRAM_NOT_FOUND: 'Program not found',
