+++ src/core_modules/capture-core/rules/actionsCreator/index.js
@@ -1,3 +1,4 @@
 // @flow
 export { getRulesActionsForEvent, getRulesActionsForTEI } from './rulesEngineActionsCreator';
-export { getCurrentClientValues, getCurrentClientMainData } from './inputHelpers';
+export { getCurrentClientValues, getCurrentClientMainData, FieldData } from './inputHelpers';
+export { actionTypes } from './rulesEngine.actions';
